+++ sentry-sidekiq/lib/sentry-sidekiq.rb
@@ -39,3 +39,6 @@ Sidekiq.configure_client do |config|
     chain.add Sentry::Sidekiq::SentryContextClientMiddleware
   end
 end
+
+# patches
+require "sentry/sidekiq/cron/job"